+++ common/src/main/java/com/niixlabs/lucidadvancements/config/category/CategoryIconType.java
@@ -1,0 +1,10 @@
+package com.niixlabs.lucidadvancements.config.category;
+
+import com.google.gson.annotations.SerializedName;
+
+public enum CategoryIconType {
+    @SerializedName("item")
+    ITEM,
+    @SerializedName("texture")
+    TEXTURE
+}